Fix bug in evaluating winning combinations
gameBoard.board is initialized with nine explicit elements (i.e. blank
strings ''), instead of relying on unmarked squares to be represented
implicitly by `undefined`. This makes checking for win conditions
simpler, requiring only a check that each mark in a group evaluates to
true(non-blank), and a check that all marks in a group are equal to each
other.

Changing the initial setup of gameBoard.board also required
corresponding changes in the methods for clearing the board and checking
if the board is empty.