Link game logic to DOM
A game can now be played by clicking on squares to make a move.

Because the game progresses by events (i.e. clicks on squares), the
game logic was moved to the display module, which communicates directly
with the gameBoard module to tie clicking on a square with checking to
see if the move ends the game. Later on, it might be bettwe if the game
logic were isolated in its own module, which would be imported by
display (display takes user input/events such as player info. and clicks
on squares, and calls the game logic to evaluate).