Indicate game state with backdrop squares
Formerly decorative backdrop squares take on semantic value.
The two squares are labeled with the markers (i.e. 'X' and 'O') of the
two players. The square corresponding to the current player is
highlighted during the turn. The winner of the last round of play is
indicated by a green stripe on the side of the corresponding square.